+++ tehreer-android/src/main/java/com/mta/tehreer/internal/util/RawUInt16List.java
@@ -17,9 +17,9 @@
 package com.mta.tehreer.internal.util;
 
 import com.mta.tehreer.internal.Raw;
-import com.mta.tehreer.util.IntList;
+import com.mta.tehreer.util.FloatList;
 
-public class RawUInt16List implements IntList {
+public class RawUInt16List implements FloatList {
 
     private final long pointer;
     private final int size;
